Multiplayer netcode, dedicated server, SIMD physics, and connect/disconnect
Map-load handshake protocol (N_MSG_MAP_LOADED / N_MSG_MAP_CONFIRMED):
- Client notifies server after loading map with FNV-1a hash validation
- Server gates snapshot broadcast on map_ready, preventing desync
- Map name included in CONNECT_ACCEPTED for remote clients
- Loopback fast-tracks handshake (same process, no round trip)

UDP transport and dedicated server:
- Fixed socket_fd type from int to intptr_t (64-bit Windows SOCKET safety)
- Real dedicated server in server_main.c with CLI args, signal handling,
  fixed timestep tick loop, and CPU-friendly sleep
- Server sends N_MSG_DISCONNECT on client removal

Mid-session join and connect/disconnect commands:
- Rising/falling edge detection on client map_ready for player connect/disconnect
- Console 'connect <ip>:<port>' command with full connection state machine
  (LOCAL -> CONNECTING -> LOADING_MAP -> HANDSHAKING -> REMOTE)
- Console 'disconnect' restores loopback gracefully
- Auto-disconnect detection with proper player state restoration
- 10-second timeout on both connection and handshake phases

SIMD optimization (SSE2/SSE4.1) for physics hot paths:
- p_simd.h: IEEE-754 compliant helpers (no rsqrt/rcp/FMA) with scalar fallbacks
- p_trace.c: SIMD dot products, Minkowski expansion, swept AABB broadphase
- p_slide.c: SIMD clip velocity and plane duplicate/re-entry checks
- p_brush.c: SIMD vertex validation and AABB overlap delegation

Code review fixes:
- Extract restore_loopback_netcode() and restore_local_gameplay() helpers,
  eliminating 6x duplicated loopback restoration boilerplate
- Fix auto-disconnect not restoring player spawn state (health/armor/weapons)
- Add 10-second timeout on CONN_MODE_HANDSHAKING (previously could hang forever)

+++ src/netcode/n_transport.c
@@ -112,7 +112,7 @@ bool n_transport_open_udp(n_transport_t *t, u16 bind_port) {
         return false;
     }
 
-    t->socket_fd = (int)sock;
+    t->socket_fd = (intptr_t)sock;
 #else
     int sock = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP);
     if (sock < 0) return false;
@@ -143,7 +143,7 @@ void n_transport_close(n_transport_t *t) {
 #ifdef QK_PLATFORM_WINDOWS
         closesocket((SOCKET)t->socket_fd);
 #else
-        close(t->socket_fd);
+        close((int)t->socket_fd);
 #endif
         t->socket_fd = -1;
     }
@@ -165,7 +165,7 @@ static i32 udp_send(n_transport_t *t, const n_address_t *to,
     if (result == SOCKET_ERROR) return -1;
     return result;
 #else
-    ssize_t result = sendto(t->socket_fd, data, len, 0,
+    ssize_t result = sendto((int)t->socket_fd, data, len, 0,
                             (struct sockaddr *)&addr, sizeof(addr));
     if (result < 0) return -1;
     return (i32)result;
@@ -184,7 +184,7 @@ static i32 udp_recv(n_transport_t *t, n_address_t *from,
     if (result == SOCKET_ERROR) return 0;
 #else
     socklen_t addr_len = sizeof(addr);
-    ssize_t result = recvfrom(t->socket_fd, data, max_len, 0,
+    ssize_t result = recvfrom((int)t->socket_fd, data, max_len, 0,
                               (struct sockaddr *)&addr, &addr_len);
     if (result < 0) return 0;
 #endif
